Move test-specific code to test file
